requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
pydantic==2.3.0
python-dotenv==1.0.0
groq==0.4.1
jinja2==3.1.2

### s:\Projects\FastAPI\SeperateFolder\MyraChatBot\.gitignore

Standard Python gitignore file:

